fix bugs with frames
initial for loop doesn't create a frame for the current time.

if a frame was made of 'no image' images and later on had actual
data, its 'no_img' property would now set to False. Before this
caused the video to run out of frames as it would ignore all the
frames that once had 'no image' images.